+++ src/Monad/Trans.hs
@@ -8,6 +8,7 @@ module Monad.Trans
        , module Control.Monad.Reader
        , module Control.Monad.State.Strict
        , module Control.Monad.Trans
+       , module Control.Monad.Trans.Identity
        , module Control.Monad.Trans.Maybe
 
          -- * Convenient functions to work with 'Reader' monad
@@ -24,16 +25,19 @@ module Monad.Trans
        ) where
 
 -- Monad transformers
-import           Control.Monad.Except       (ExceptT (..), runExceptT)
-import           Control.Monad.Reader       (MonadReader, Reader, ReaderT (..), ask, asks,
-                                             local, reader, runReader)
-import           Control.Monad.State.Strict (MonadState, State, StateT (..), evalState,
-                                             evalStateT, execState, execStateT, gets,
-                                             modify, runState, state, withState)
-import           Control.Monad.Trans        (MonadIO, MonadTrans, lift, liftIO)
-import           Control.Monad.Trans.Maybe  (MaybeT (..), exceptToMaybeT, maybeToExceptT)
+import           Control.Monad.Except         (ExceptT (..), runExceptT)
+import           Control.Monad.Reader         (MonadReader, Reader, ReaderT (..), ask,
+                                               asks, local, reader, runReader)
+import           Control.Monad.State.Strict   (MonadState, State, StateT (..), evalState,
+                                               evalStateT, execState, execStateT, get,
+                                               gets, modify, put, runState, state,
+                                               withState)
+import           Control.Monad.Trans          (MonadIO, MonadTrans, lift, liftIO)
+import           Control.Monad.Trans.Identity (IdentityT (runIdentityT))
+import           Control.Monad.Trans.Maybe    (MaybeT (..), exceptToMaybeT,
+                                               maybeToExceptT)
 
-import           Prelude                    (Functor, flip, fst, snd, (<$>))
+import           Prelude                      (Functor, flip, fst, snd, (<$>))
 
 -- | Shorter and more readable alias for @flip runReaderT@.
 usingReaderT :: r -> ReaderT r m a -> m a